+++ src/components/editor/adapter/helpers.ts
@@ -1,0 +1,16 @@
+// created with chatGPT
+export function createObjectFromPath<T>(path: string, value: T): any {
+  const keys = path.split('.');
+  const result: any = {};
+  let current = result;
+  for (let i = 0; i < keys.length; i++) {
+    const key = keys[i];
+    if (i === keys.length - 1) {
+      current[key] = value;
+    } else {
+      current[key] = {};
+      current = current[key];
+    }
+  }
+  return result;
+}
