+++ src/components/editor/adapter/input-text-adapter.tsx
@@ -1,0 +1,38 @@
+import * as React from 'react';
+import { useSelector } from 'react-redux';
+import { apiState } from '@/state/api';
+import { InputText } from '@/components/editor/form';
+
+import { Field } from '@/shared/types';
+import { createObjectFromPath } from './helpers';
+
+const { useProfileUpdater } = apiState.profile;
+
+export type InputTextAdapterProps = {
+  path: Field['path']
+}
+
+export function InputTextAdapter(props: InputTextAdapterProps) {
+  const value = useSelector(apiState.resume.selectors.selectResumeProperty(props.path, ''));
+  const update = useProfileUpdater();
+
+  const handleChange = React.useCallback(
+    (value: string) => {
+      update(createObjectFromPath(props.path, value));
+    },
+    [update, props.path]
+  );
+
+  if (typeof value !== 'string') {
+    console.error('Invalid value type in InputTextAdapter');
+    return null;
+  }
+
+  return (
+    <InputText
+      label="Title"
+      value={value}
+      onChange={handleChange}
+    />
+  );
+}
