+++ src/components/editor/common/index.ts
@@ -4,3 +4,4 @@ export * from './inline';
 export * from './text';
 export * from './box';
 export * from './selectable-card';
+export * from './adapter';
